% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_row_if.R
\name{drop_row_if}
\alias{drop_row_if}
\title{Conditionally drop rows based on percent missingness}
\usage{
drop_row_if(df, sign = "gt", percent_na = 20, as_percent = TRUE)
}
\arguments{
\item{df}{A data.frame object}

\item{sign}{Character. One of gteq,lteq,lt,gt or eq which refer to greater than(gt) or equal(eq) or
less than(lt) or equal to(eq) respectively.}

\item{percent_na}{The percentage to use when dropping columns with missing values}

\item{as_percent}{Logical. If set to TRUE, percent_na is treated as a percentage. Otherwise, decimals(fractions) are 
used.}
}
\description{
Conditionally drop rows based on percent missingness
}
\examples{
head(drop_row_if(airquality, sign="gteq", percent_na = 50))
# should give the same output as above. 
head(drop_row_if(airquality, sign="gteq", percent_na = 0.5, as_percent=FALSE))
}
