% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_na_means}
\alias{get_na_means}
\title{Get mean missingness.}
\usage{
get_na_means(x, as_percent = TRUE)
}
\arguments{
\item{x}{A vector whose mean NA is required.}

\item{as_percent}{Boolean? Report means as percents, defaults to TRUE.}
}
\description{
Get mean missingness.
}
\examples{
get_na_means(airquality)
}
