% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeWorld.R
\name{MakeWorld}
\alias{MakeWorld}
\title{Create a simple domain raster}
\usage{
MakeWorld(x.dim, y.dim, initialValue = 0)
}
\arguments{
\item{x.dim}{Dimension for the x-axis, in pixels}

\item{y.dim}{Dimension for the y-axis, in pixels}

\item{initialValue}{Value for initializing the raster cell values.  Defaults to 0. Pixels with NA values are interpreted as out of the domain by the \code{SpreadingDye::GrowRange()} and \code{SpreadingDye::SDMDE()} functions}
}
\description{
This function creates simple rectangular domain rasters, for use with the MDE function.
}
\examples{
# a 10 by 10 raster with NA for each cell value
MakeWorld(10, 10)

# a 20 by 10 raster with 1 for each cell value
MakeWorld(20, 10, initialValue = 1)
}
