% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mde.R
\name{mde}
\alias{mde}
\title{Finding mode of a variable}
\usage{
mde(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector containing values of a variable .}

\item{na.rm}{default value to include NA values. It can be set to TRUE to exclude NA values.}
}
\value{
a list [1] modal values [2] number of modes in the variable .
}
\description{
This function find the mode of a categorical variable . If the variable is not
categorical it displays a warning message .
}
\details{
This function find the mode of a categorical variable. It check whether there are
NA values in the vector. There is option to include or exclude the NA values.
Default is na.rm=FALSE . If the variable is not categorical it displays a warning
message.
}
\examples{
x <- c(2,3,4,4,7,4,9)
xm <- mode(x)
}
\author{
Shan
}
\seealso{
\code{unique}      \code{is.na}
}

