% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{na_counts}
\alias{na_counts}
\title{Get NA counts for a given character, numeric, factor, etc.}
\usage{
na_counts(x)
}
\arguments{
\item{x}{A vector whose number of missing values is to be determined.}
}
\description{
Get NA counts for a given character, numeric, factor, etc.
}
\examples{
na_counts(airquality$Ozone) 
}
