% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_missing.R
\name{percent_na}
\alias{percent_na}
\title{percent missing but for vectors.}
\usage{
percent_na(x)
}
\arguments{
\item{x}{A vector whose mean NA is required.}
}
\description{
percent missing but for vectors.
}
\examples{
percent_na(airquality$Ozone)
}
