% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{recode_helper}
\alias{recode_helper}
\title{Helper functions in package mde}
\usage{
recode_helper(
  x,
  pattern_type = NULL,
  pattern = NULL,
  original_value,
  new_value,
  case_sensitive = FALSE,
  ...
)
}
\arguments{
\item{x}{A data.frame object}

\item{pattern_type}{One of contains', 'starts_with' or 'ends_with'.}

\item{pattern}{A character pattern to match}

\item{original_value}{Value to replace}

\item{new_value}{Replacement value.}

\item{case_sensitive}{Defaults to FALSE. Patterns are case insensitive if TRUE}

\item{...}{Other arguments to other functions}
}
\description{
Helper functions in package mde
}
