% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{recode_selectors}
\alias{recode_selectors}
\title{Helpedr functions in package mde}
\usage{
recode_selectors(
  x,
  pattern_type = NULL,
  pattern = NULL,
  case_sensitive = FALSE,
  ...
)
}
\arguments{
\item{x}{A data.frame object}

\item{pattern_type}{One of contains', 'starts_with' or 'ends_with'.}

\item{pattern}{A character pattern to match}

\item{case_sensitive}{Defaults to FALSE. Patterns are case insensitive if TRUE}

\item{...}{Other arguments to other functions}
}
\description{
Helpedr functions in package mde
}
