% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_na_at.R
\name{drop_na_at}
\alias{drop_na_at}
\title{Drop missing values at columns that match a given pattern}
\usage{
drop_na_at(x, pattern_type = "contains", pattern = NULL, ...)
}
\arguments{
\item{x}{A data.frame object}

\item{pattern_type}{One of "contains", "ends_with" or "starts_with"}

\item{pattern}{The type of pattern to use when matching the pattern_type. The pattern is case sensitive}

\item{...}{Other params to other methods}
}
\value{
A data.frame object containing only columns that match the given pattern with the missing values removed.
}
\description{
Provides a simple yet efficient way to drop missing values("NA"s) at
columns that match a given pattern.
}
\examples{
head(drop_na_at(airquality,pattern_type = "starts_with","O"))
}
