% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_na_if.R
\name{drop_na_if}
\alias{drop_na_if}
\title{Conditon based dropping of columns with missing values}
\usage{
drop_na_if(x, percent_na = 50, ...)
}
\arguments{
\item{x}{A data.frame object}

\item{percent_na}{The percentage to use when dropping columns with missing values}

\item{...}{Other arguments to "percent_missing"}
}
\value{
A data.frame object with columns  that meet the target criteria dropped
}
\description{
"drop_na_if" provides a simple way to drop columns with missing values if
they meet certain criteria/conditions.
}
\details{
When dropping columns, percent_na is taken as the minimum percentage ie any columns
with greater than or equal to percent_na missingness are dropped(currently)
}
\examples{
drop_na_if(airquality, percent_na = 24)
}
\seealso{
\code{\link{percent_missing}}
}
