% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_na_counts.R
\name{get_na_counts}
\alias{get_na_counts}
\title{Add columnwise counts of missing values}
\usage{
get_na_counts(x)
}
\arguments{
\item{x}{A valid R `object` for which `na_counts` are needed.}
}
\value{
An object of the same type as `x` showing the respective number of missing values.
}
\description{
This function takes a `data.frame` object as an input and returns the corresponding
`NA` counts. `NA` refers to R's builtin missing data holder.
}
\examples{
get_na_counts(airquality)
}
