% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_summary.R
\name{na_summary}
\alias{na_summary}
\title{An all-in-one missingness report}
\usage{
na_summary(
  df,
  grouping_cols = NULL,
  sort_by = NULL,
  descending = FALSE,
  exclude_cols = NULL,
  include_pattern_type = NULL,
  include_pattern = NULL,
  exclude_pattern_type = NULL,
  exclude_pattern = NULL,
  round_to = NULL
)
}
\arguments{
\item{df}{A valid R `object` for which the percentage of missing values is 
required.}

\item{grouping_cols}{A character vector. If supplied, one can provide the columns by which to group the
data.}

\item{sort_by}{One of counts or percents. This determines whether the results are sorted by counts or percentages.}

\item{descending}{Logical. Should missing values be sorted in decreasing order ie largest to smallest? Defaults to FALSE.}

\item{exclude_cols}{A character vector indicating columns to exclude when 
returning results.}

\item{include_pattern_type}{A regular expression type. One of "starts_with",
"contains", or "regex". Defaults to NULL. Only use for selective inclusion.}

\item{include_pattern}{A character specifying the pattern to use as the}

\item{exclude_pattern_type}{A regular expression type. for exclusion.}

\item{exclude_pattern}{Pattern to use for exclusion. 
column inclusion criteria.}

\item{round_to}{Number of places to round 2. Defaults to user digits option.}
}
\description{
An all-in-one missingness report
}
\examples{
na_summary(airquality)
# grouping
test2 <- data.frame(ID= c("A","A","B","A","B"),Vals = c(rep(NA,4),"No"),
ID2 = c("E","E","D","E","D"))
na_summary(test2,grouping_cols = c("ID","ID2"))
# sort summary
na_summary(airquality,sort_by = "percent_missing",descending = TRUE)
na_summary(airquality,sort_by = "percent_complete")
# Include only via a regular expression
na_summary(mtcars, include_pattern_type = "contains", 
include_pattern = "mpg|disp|wt")
na_summary(airquality, include_pattern_type = "starts_with", 
include_pattern = "ozone")
# exclusion via a regex 
na_summary(airquality, exclude_pattern_type = "starts_with",
exclude_pattern = "oz|Sol")
}
