% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_summary.R
\name{na_summary}
\alias{na_summary}
\title{An all-in-one missingness report}
\usage{
na_summary(df, ...)
}
\arguments{
\item{df}{A data.frame object}

\item{...}{Other params to other methods}
}
\description{
An all-in-one missingness report
}
\examples{
na_summary(airquality)
}
