% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_summary.R
\name{na_summary}
\alias{na_summary}
\title{An all-in-one missingness report}
\usage{
na_summary(df, grouping_cols = NULL, sort_by = NULL, descending = FALSE, ...)
}
\arguments{
\item{df}{A valid R `object` for which the percentage of missing values is required.}

\item{grouping_cols}{A character vector. If supplied, one can provide the columns by which to group the
data.}

\item{sort_by}{One of counts or percents. This determines whether the results are sorted by counts or percentages.}

\item{descending}{Logical. Should missing values be sorted in decreasing order ie largest to smallest? Defaults to FALSE.}

\item{...}{Arguments to other functions}
}
\description{
An all-in-one missingness report
}
\examples{
na_summary(airquality)
# grouping
test2 <- data.frame(ID= c("A","A","B","A","B"),
Vals = c(rep(NA,4),"No"),ID2 = c("E","E","D","E","D"))
na_summary(test2,grouping_cols = c("ID","ID2"))


}
