% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_missing.R
\name{percent_missing}
\alias{percent_missing}
\title{Columnwise missingness percentages}
\usage{
percent_missing(x, percent = TRUE)
}
\arguments{
\item{x}{A valid R `object` for which the percentage of missing values is required.}

\item{percent}{Logical. Should values be displayed as percentages? Defaults to `TRUE`, otherwise
results are shown in decimals.}
}
\value{
An object of the same class as x showing the percentage of missing values.
}
\description{
A convenient way to obtain percent missingness columnwise.
}
\examples{
percent_missing(airquality)
# show decimals
percent_missing(airquality, percent = FALSE)
}
