% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_na_as.R
\name{recode_na_as}
\alias{recode_na_as}
\title{Replace missing values with another value}
\usage{
recode_na_as(x, value = 0, subset_df = FALSE, subset_cols = NULL)
}
\arguments{
\item{x}{A valid R `object` for which `na_counts` are needed.}

\item{value}{The value to convert to `NA`. We can for instance change "n/a" to `NA` or any other value.}

\item{subset_df}{Logical. Use only specific columns? Defaults to FALSE. All "value"s everywhere are "recoded".}

\item{subset_cols}{Vector. If subset_df is TRUE, then this provides the columns for which changes are required.}
}
\value{
An object of the same type as x with NAs replaced with the desired value.
}
\description{
This provides a convenient way to recode "NA" as another value for instance "NaN", "n/a" or
any other value a user wishes to use.
}
\examples{
recode_na_as(airquality, "n/a")
recode_na_as(airquality, subset_df = TRUE,
subset_cols = "Ozone", value = "N/A")
}
